Initial Implementation of Sound Corr. Output
This is going to need to be rethought a bit before it works properly. The program currently only outputs ID pairs for sounds for which a match could be found; so using these output files would leave some SFX IDs unchanged, likely causing a crash.

There's also a rather important design decision to be made in addressing what a given unmatched sound's SFX ID should be mapped to. Tentatively I'd map it to se_001, but I'd need to make it really clear to users that such a change had taken place so they can fix it when sounds play unexpectedly.